# backend/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AIInterviewCoach.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/submissions": {
      "post": {
        "tags": [
          "Submissions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/submissions/session/{interviewSessionId}": {
      "get": {
        "tags": [
          "Submissions"
        ],
        "parameters": [
          {
            "name": "interviewSessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Submissions"
        ],
        "parameters": [
          {
            "name": "interviewSessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/submissions/problem/{problemId}": {
      "delete": {
        "tags": [
          "Submissions"
        ],
        "parameters": [
          {
            "name": "problemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "interviewSessionId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateSubmissionRequestDto": {
        "type": "object",
        "properties": {
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "interviewSessionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "sourceCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SubmissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "interviewSessionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "sourceCode": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "passedTests": {
            "type": "integer",
            "format": "int32"
          },
          "totalTests": {
            "type": "integer",
            "format": "int32"
          },
          "executionTimeMs": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "memoryKb": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "executionOutput": {
            "type": "string",
            "nullable": true
          },
          "aiFeedback": {
            "type": "string",
            "nullable": true
          },
          "aiFeedbackSource": {
            "type": "string",
            "nullable": true
          },
          "aiFeedbackStatus": {
            "type": "string",
            "nullable": true
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}